roll county/tract source components into layer components
